Split widget: signal detail
